+++ internal/ui/fs.go
@@ -4,5 +4,5 @@ import "embed"
 
 // DistFS contains the embedded frontend assets from the dist directory.
 //
-//go:embed dist/*
+//go:embed dist
 var DistFS embed.FS